Moved models from service package to model package.
