+++ rest/src/main/java/restwars/rest/resources/ResearchSubResource.java
@@ -7,13 +7,13 @@ import com.wordnik.swagger.annotations.ApiParam;
 import io.dropwizard.auth.Auth;
 import restwars.model.planet.Planet;
 import restwars.model.player.Player;
+import restwars.model.technology.Research;
 import restwars.model.technology.TechnologyType;
 import restwars.rest.mapper.ResearchMapper;
 import restwars.rest.resources.param.LocationParam;
 import restwars.restapi.dto.technology.ResearchRequest;
 import restwars.restapi.dto.technology.ResearchResponse;
 import restwars.service.planet.PlanetService;
-import restwars.service.technology.Research;
 import restwars.service.technology.ResearchException;
 import restwars.service.technology.TechnologyService;
 import restwars.util.Functional;
