+++ service/src/main/java/restwars/model/technology/Research.java
@@ -1,8 +1,7 @@
-package restwars.service.technology;
+package restwars.model.technology;
 
 import com.google.common.base.Objects;
 import com.google.common.base.Preconditions;
-import restwars.model.technology.TechnologyType;
 
 import java.util.UUID;
