+++ service/src/main/java/restwars/service/flight/impl/FlightServiceImpl.java
@@ -11,6 +11,7 @@ import restwars.model.planet.Location;
 import restwars.model.planet.Planet;
 import restwars.model.player.Player;
 import restwars.model.resource.Resources;
+import restwars.model.ship.Hangar;
 import restwars.model.ship.Ship;
 import restwars.model.ship.ShipType;
 import restwars.model.ship.Ships;
@@ -34,7 +35,6 @@ import restwars.service.mechanics.PlanetMechanics;
 import restwars.service.mechanics.ShipMechanics;
 import restwars.service.mechanics.TechnologyMechanics;
 import restwars.service.planet.PlanetDAO;
-import restwars.service.ship.Hangar;
 import restwars.service.ship.HangarDAO;
 import restwars.service.ship.impl.ShipUtils;
 import restwars.service.technology.TechnologyDAO;
