+++ service/src/main/java/restwars/service/flight/impl/flighthandler/AbstractFlightHandler.java
@@ -6,6 +6,7 @@ import org.slf4j.LoggerFactory;
 import restwars.model.flight.Flight;
 import restwars.model.flight.FlightDirection;
 import restwars.model.resource.Resources;
+import restwars.model.ship.Hangar;
 import restwars.model.ship.Ships;
 import restwars.service.event.EventService;
 import restwars.service.flight.DetectedFlightDAO;
@@ -14,7 +15,6 @@ import restwars.service.infrastructure.RoundService;
 import restwars.service.infrastructure.UUIDFactory;
 import restwars.service.mechanics.ShipMechanics;
 import restwars.service.planet.PlanetDAO;
-import restwars.service.ship.Hangar;
 import restwars.service.ship.HangarDAO;
 import restwars.service.ship.impl.ShipUtils;
 import restwars.util.MathExt;
