+++ service/src/main/java/restwars/service/flight/impl/flighthandler/AttackFlightHandler.java
@@ -7,6 +7,7 @@ import restwars.model.fight.Fight;
 import restwars.model.flight.Flight;
 import restwars.model.planet.Planet;
 import restwars.model.resource.Resources;
+import restwars.model.ship.Hangar;
 import restwars.model.ship.Ships;
 import restwars.service.event.EventService;
 import restwars.service.fight.FightDAO;
@@ -17,7 +18,6 @@ import restwars.service.infrastructure.RoundService;
 import restwars.service.infrastructure.UUIDFactory;
 import restwars.service.mechanics.ShipMechanics;
 import restwars.service.planet.PlanetDAO;
-import restwars.service.ship.Hangar;
 import restwars.service.ship.HangarDAO;
 
 import java.util.Optional;
