+++ service/src/main/java/restwars/service/flight/impl/flighthandler/TransferFlightHandler.java
@@ -5,6 +5,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import restwars.model.flight.Flight;
 import restwars.model.planet.Planet;
+import restwars.model.ship.Hangar;
 import restwars.service.event.EventService;
 import restwars.service.flight.DetectedFlightDAO;
 import restwars.service.flight.FlightDAO;
@@ -12,7 +13,6 @@ import restwars.service.infrastructure.RoundService;
 import restwars.service.infrastructure.UUIDFactory;
 import restwars.service.mechanics.ShipMechanics;
 import restwars.service.planet.PlanetDAO;
-import restwars.service.ship.Hangar;
 import restwars.service.ship.HangarDAO;
 
 import java.util.Optional;
