+++ service/src/main/java/restwars/service/ship/HangarDAO.java
@@ -1,5 +1,7 @@
 package restwars.service.ship;
 
+import restwars.model.ship.Hangar;
+
 import java.util.Optional;
 import java.util.UUID;
