+++ service/src/main/java/restwars/service/technology/impl/TechnologyServiceImpl.java
@@ -8,6 +8,7 @@ import restwars.model.building.Buildings;
 import restwars.model.planet.Planet;
 import restwars.model.player.Player;
 import restwars.model.resource.Resources;
+import restwars.model.technology.Research;
 import restwars.model.technology.Technologies;
 import restwars.model.technology.Technology;
 import restwars.model.technology.TechnologyType;
@@ -19,7 +20,10 @@ import restwars.service.infrastructure.UUIDFactory;
 import restwars.service.mechanics.BuildingMechanics;
 import restwars.service.mechanics.TechnologyMechanics;
 import restwars.service.planet.PlanetDAO;
-import restwars.service.technology.*;
+import restwars.service.technology.ResearchDAO;
+import restwars.service.technology.ResearchException;
+import restwars.service.technology.TechnologyDAO;
+import restwars.service.technology.TechnologyService;
 import restwars.util.MathExt;
 
 import javax.inject.Inject;
