+++ storage/src/main/java/restwars/storage/ship/JooqHangarDAO.java
@@ -6,10 +6,10 @@ import org.jooq.Record;
 import org.jooq.Result;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import restwars.model.ship.Hangar;
 import restwars.model.ship.Ship;
 import restwars.model.ship.ShipType;
 import restwars.model.ship.Ships;
-import restwars.service.ship.Hangar;
 import restwars.service.ship.HangarDAO;
 import restwars.service.unitofwork.UnitOfWorkService;
 import restwars.storage.jooq.AbstractJooqDAO;
