+++ storage/src/main/java/restwars/storage/technology/JooqResearchDAO.java
@@ -4,8 +4,8 @@ import com.google.common.base.Preconditions;
 import org.jooq.Result;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import restwars.model.technology.Research;
 import restwars.model.technology.TechnologyType;
-import restwars.service.technology.Research;
 import restwars.service.technology.ResearchDAO;
 import restwars.service.unitofwork.UnitOfWorkService;
 import restwars.storage.jooq.AbstractJooqDAO;
